renamed config namespace to configuration
