+++ ArchitectsLibrary/API/AchievementServices.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
-using ArchitectsLibrary.Config;
+using ArchitectsLibrary.Configuration;
 
 namespace ArchitectsLibrary.API
 {
