+++ ArchitectsLibrary/ArchitectsLibrary.csproj
@@ -154,7 +154,8 @@
     <Compile Include="Buildables\BuildableSpecimenCases.cs" />
     <Compile Include="Buildables\BuildableTable.cs" />
     <Compile Include="Buildables\BuildableWarper.cs" />
-    <Compile Include="Config\AchievementData.cs" />
+    <Compile Include="Configuration\AchievementData.cs" />
+    <Compile Include="Configuration\Config.cs" />
     <Compile Include="CreatorKit\Mono\FilterableList.cs" />
     <Compile Include="CreatorKit\Mono\FilterableListItem.cs" />
     <Compile Include="CreatorKit\Mono\GenericButton.cs" />
